requirements.txt
streamlit>=1.28.0
openai>=1.0.0
pillow>=9.0.0
icalendar>=5.0.0
pyyaml>=6.0
pathlib2>=2.3.0
numpy>=1.23.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
tensorflow>=2.13.0
yfinance>=0.2.31
ta>=0.10.0
plotly>=5.15.0
